mm: process_reclaim: Modified the driver to kill processes.
This is marked HIGHLY EXPERIMENTAL. Use at your own risk.

If you want to test this, ZRAM/SWAP must be enabled!

Background
==========
While playing around with various low memory solutions on Android, like
the stock Android LowMemoryKiller and Sultan's Simple LMK,
I was unsatisfied with the way with which they handled low memory.

Here are my reasons,but I may be wrong/unclear.Please do feel free to correct me.

- Android LMK is too aggressive,but setting minfrees/adjs help to an
  extent.During my testing,I found that LMK depletes swap-pages to 0,
  and a thrashing situation ensues.When this thrashing situation
  continues for a long time,OOM Killer comes into play,and starts
  killing tasks at random.This is undesirable.

- Sultan's SLMK is a bit better,but still aggressive,in the sense that
  it starts to kill tasks even if there is a short memory spike.This is
  undesirable, as there may be a situation wherein many swap pages are
  available,but since tasks are being killed when those memory spikes
  occur,swap pages may not be used as much as they are freed.These kind
  of situations are frequent when using SLMK during my testing.

In short,I was unable to keep many apps open as per the potential of my
device's RAM and ZRAM.The above-mentioned solutions also killed Android
services at random, which proved to be annoying, for example when I had a
music player and apps in the background,the music player would abruptly be
killed.

My objective is to:

- Keep important tasks like Android services running
  until they are closed by the user.

- Kill apps as little as possible,or specifically,killing apps based
  on the total time spent on it.This means that, apps with larger usage
  time get killed more rarely,and apps with lesser usage times get
  killed frequently.

Solution
========
This is my humble(flawed?) attempt at handling a low-memory situation in a
better manner.

I used the Process Reclaim driver as the base for this purpose.
The reason why I used this driver was that,its function was to reclaim
ANONYMOUS(MM_ANON) pages from each scanned task as much as possible(according to per_swap_size),
so that apps never get killed,and memory gets reclaimed in the end.

In theory,this sounds good.While testing though,I ran into a very
serious problem: system freezes.Too much memory pressure and less ANONYMOUS
pages to reclaim rendered the system unusable for a while,which was NOT cool.

Therefore,the only way to prevent these system freezes was to kill some
tasks,but at the appropriate time.This is the approach I took:

- We keep reclaiming pages if they are above defined threshold
  (free_swap_limit).

- Once the swap goes below a percentage(free_percent) above/below the threshold,
  we start to collect tasks to kill.

- We then sort those tasks based on their last accessed (stime+utime),
  and then start killing the tasks with the lowest (stime+utime).We stop
  killing tasks once the number of swap pages are above the threshold
  (free_swap_limit).
  I specifically sorted according to acct_timexpd,see kernel/tsacct.c for
  more details(CONFIG_TASK_XACCT).

So, in theory, this should reclaim ANONYMOUS pages more often, and kill
tasks less,depending on the memory pressure.

Other Changes
=============
- I've renamed some variables in the driver, so that their values don't
  get reset by the values set by the ROM, specifically,
  init.qcom.post_boot.sh

- Removed min_score_adj, I've now added a blacklist of OOM_SCORE_ADJ's
  so that the ones in the blacklist don't get their MM_ANON pages
  reclaimed.This is important,because some running services/apps may
  glitch(this happens rarely).For example, while playing music, I
  experienced crackling sounds at random intervals.

- Added functionality to wake up oom_reaper after killing a task,so that
  once the task is killed,all its pages(?) will be reaped.
  (Taken from the lowmemorykiller.c code)

- Added some code bits from SLMK(marked in the comments).

- Removed traces from the driver.

Credits
=======
Minchan Kim and Vinayak Menon for the Process Reclaim driver, and
Sultan AlSawaf(kerneltoast) for the appropriate code bits.

Signed-off-by: darkhz <kmachanwenw@gmail.com>
Signed-off-by: Alex Winkowski <dereference23@outlook.com>